+++ src/main/java/com/app/kokonut/company/CompanyRepository.java
@@ -11,7 +11,7 @@ import org.springframework.stereotype.Repository;
  * Remark :
  */
 @Repository
-public interface CompanyRepository extends JpaRepository<Company, Integer>, JpaSpecificationExecutor<Company>, CompanyRepositoryCustom {
+public interface CompanyRepository extends JpaRepository<Company, Long>, JpaSpecificationExecutor<Company>, CompanyRepositoryCustom {
 
     boolean existsByBusinessNumber(String businessNumber);
