+++ src/main/java/com/app/kokonut/email/emailGroup/EmailGroupRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
 
-public interface EmailGroupRepository extends JpaRepository<EmailGroup, Integer>, JpaSpecificationExecutor<EmailGroup>, EmailGroupRepositoryCustom {
+public interface EmailGroupRepository extends JpaRepository<EmailGroup, Long>, JpaSpecificationExecutor<EmailGroup>, EmailGroupRepositoryCustom {
 
 
 }
