+++ src/main/java/com/app/kokonut/email/emailHistory/EmailHistoryRepository.java
@@ -3,6 +3,6 @@ package com.app.kokonut.email.emailHistory;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface EmailHistoryRepository extends JpaRepository<EmailHistory, Integer>, JpaSpecificationExecutor<EmailHistory> {
+public interface EmailHistoryRepository extends JpaRepository<EmailHistory, Long>, JpaSpecificationExecutor<EmailHistory> {
 
 }
