+++ src/main/java/com/app/kokonut/faq/FaqRepository.java
@@ -3,6 +3,6 @@ package com.app.kokonut.faq;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface FaqRepository extends JpaRepository<Faq, Integer>, JpaSpecificationExecutor<Faq>, FaqRepositoryCustom {
+public interface FaqRepository extends JpaRepository<Faq, Long>, JpaSpecificationExecutor<Faq>, FaqRepositoryCustom {
 
 }
