+++ src/main/java/com/app/kokonut/refactor/adminRemove/AdminRemoveRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AdminRemoveRepository extends JpaRepository<AdminRemove, Integer>, JpaSpecificationExecutor<AdminRemove>, AdminRemoveRepositoryCustom {
+public interface AdminRemoveRepository extends JpaRepository<AdminRemove, Long>, JpaSpecificationExecutor<AdminRemove>, AdminRemoveRepositoryCustom {
 
 }
