+++ src/main/java/com/app/kokonut/refactor/file/repository/FileRepository.java
@@ -4,6 +4,6 @@ import com.app.kokonut.refactor.file.entity.File;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface FileRepository extends JpaRepository<File, Integer>, JpaSpecificationExecutor<File> {
+public interface FileRepository extends JpaRepository<File, Long>, JpaSpecificationExecutor<File> {
 
 }
