+++ src/main/java/com/app/kokonut/refactor/policy/repository/PolicyRepository.java
@@ -4,6 +4,6 @@ import com.app.kokonut.refactor.policy.entity.Policy;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface PolicyRepository extends JpaRepository<Policy, Integer>, JpaSpecificationExecutor<Policy> {
+public interface PolicyRepository extends JpaRepository<Policy, Long>, JpaSpecificationExecutor<Policy> {
 
 }
