+++ src/main/java/com/app/kokonut/refactor/privacyEmail/repository/PrivacyEmailRepository.java
@@ -4,6 +4,6 @@ import com.app.kokonut.refactor.privacyEmail.entity.PrivacyEmail;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface PrivacyEmailRepository extends JpaRepository<PrivacyEmail, Integer>, JpaSpecificationExecutor<PrivacyEmail> {
+public interface PrivacyEmailRepository extends JpaRepository<PrivacyEmail, Long>, JpaSpecificationExecutor<PrivacyEmail> {
 
 }
