+++ src/main/java/com/app/kokonut/refactor/privacyEmailHistory/repository/PrivacyEmailHistoryRepository.java
@@ -4,6 +4,6 @@ import com.app.kokonut.refactor.privacyEmailHistory.entity.PrivacyEmailHistory;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface PrivacyEmailHistoryRepository extends JpaRepository<PrivacyEmailHistory, Integer>, JpaSpecificationExecutor<PrivacyEmailHistory> {
+public interface PrivacyEmailHistoryRepository extends JpaRepository<PrivacyEmailHistory, Long>, JpaSpecificationExecutor<PrivacyEmailHistory> {
 
 }
