+++ src/main/java/com/app/kokonut/refactor/statisticsDay/StatisticsDayRepository.java
@@ -3,6 +3,6 @@ package com.app.kokonut.refactor.statisticsDay;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface StatisticsDayRepository extends JpaRepository<StatisticsDay, Integer>, JpaSpecificationExecutor<StatisticsDay> {
+public interface StatisticsDayRepository extends JpaRepository<StatisticsDay, Long>, JpaSpecificationExecutor<StatisticsDay> {
 
 }
