+++ src/main/java/com/app/kokonut/refactor/statisticsDaySystem/StatisticsDaySystemRepository.java
@@ -3,6 +3,6 @@ package com.app.kokonut.refactor.statisticsDaySystem;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface StatisticsDaySystemRepository extends JpaRepository<StatisticsDaySystem, Integer>, JpaSpecificationExecutor<StatisticsDaySystem> {
+public interface StatisticsDaySystemRepository extends JpaRepository<StatisticsDaySystem, Long>, JpaSpecificationExecutor<StatisticsDaySystem> {
 
 }
