+++ src/main/java/com/app/kokonut/revisedDocumentFile/RevisedDocumentFileRepository.java
@@ -3,5 +3,5 @@ package com.app.kokonut.revisedDocumentFile;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface RevisedDocumentFileRepository extends JpaRepository<RevisedDocumentFile, Integer> , JpaSpecificationExecutor<RevisedDocumentFile> {
+public interface RevisedDocumentFileRepository extends JpaRepository<RevisedDocumentFile, Long> , JpaSpecificationExecutor<RevisedDocumentFile> {
 }
